fix(clippy): resolve all remaining clippy warnings
- Fix str::replace chains (use replace([' ', '-'], "_"))
- Fix strip_prefix usage for ~ paths
- Fix chars_next_cmp (use starts_with)
- Fix manual_range_contains (use range.contains())
- Fix single_char_add_str (use push(char))
- Fix unnecessary casts and lifetime elision
- Fix unused imports, variables, fields
- Add Default impls for Layer2, Layer3
- Add #[allow(dead_code)] for unused Config functions
- Fix serde_json::json import in test scope
- Fix unneeded struct pattern in CLI test

+++ crates/core/src/entity_detector.rs
@@ -697,6 +697,7 @@ pub struct DetectionResult {
 
 #[derive(Debug, Clone)]
 struct ScoredEntity {
+    #[allow(unused)]
     name: String,
     person_score: i32,
     project_score: i32,
@@ -816,7 +817,7 @@ fn score_entity(name: &str, text: &str, lines: &[&str]) -> ScoredEntity {
         }
     }
     if pronoun_hits > 0 {
-        person_score += (pronoun_hits * 2) as i32;
+        person_score += pronoun_hits * 2;
         person_signals.push(format!("pronoun nearby ({}x)", pronoun_hits));
     }
